% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_quarto_ext.R
\name{use_quarto_ext}
\alias{use_quarto_ext}
\title{Install bundled Quarto extensions into current working directory}
\usage{
use_quarto_ext(ext_name = "grouped-tabsets")
}
\arguments{
\item{ext_name}{String indicating which extension to install}
}
\value{
a message if extension was successfully copied over
}
\description{
Install bundled Quarto extensions into current working directory
}
